Refactor blobs into a separate header, add a little documentation, and make better error reporting if the dataset file is not found.
